+++ R/samples_netmeta.R
@@ -1,0 +1,39 @@
+samples_netmeta <- function(x, nsim, pooled, small.values) {
+  
+  chkclass(x, "netmeta")
+  #
+  pooled <- setchar(pooled, c("common", "random"))
+  small.values <- setsv(small.values)
+  #
+  if (pooled == "common") {
+    TE.pooled <- x$TE.common
+    Cov.pooled <- x$Cov.common
+  }
+  else {
+    TE.pooled <- x$TE.random
+    Cov.pooled <- x$Cov.random
+  }
+  #
+  if (small.values == "desirable")
+    theta <- TE.pooled[, 1]
+  else
+    theta <- TE.pooled[1, ]
+  
+  compMatrix <- matrix(0, nrow = nrow(Cov.pooled), ncol = length(x$trts))
+  rownames(compMatrix) <- rownames(Cov.pooled)
+  colnames(compMatrix) <- x$trts
+  #
+  allcomps <- compsplit(rownames(compMatrix), x$sep.trts)
+  for (i in seq_len(nrow(Cov.pooled)))
+    compMatrix[i, allcomps[[i]]] <- 1
+  #
+  var.theta <- as.vector(ginv(compMatrix) %*% diag(Cov.pooled))
+  
+  samples <- mvtnorm::rmvnorm(nsim, theta, diag(var.theta))
+  rownames(samples) <- seq_len(nrow(samples))
+  colnames(samples) <- x$trts
+  
+  res <- list(samples = samples, compMatrix = compMatrix)
+  #
+  res
+}